+++ pdata/internal/json/number_test.go
@@ -232,3 +232,56 @@ func TestReadUint64(t *testing.T) {
 		})
 	}
 }
+
+func TestReadFloat64(t *testing.T) {
+	tests := []struct {
+		name    string
+		jsonStr string
+		want    float64
+		wantErr bool
+	}{
+		{
+			name:    "number",
+			jsonStr: `3.14 `,
+			want:    3.14,
+		},
+		{
+			name:    "string",
+			jsonStr: `"3.14"`,
+			want:    3.14,
+		},
+		{
+			name:    "negative number",
+			jsonStr: `-3.14 `,
+			want:    -3.14,
+		},
+		{
+			name:    "negative string",
+			jsonStr: `"-3.14"`,
+			want:    -3.14,
+		},
+		{
+			name:    "wrong string",
+			jsonStr: `"3.f14"`,
+			wantErr: true,
+		},
+		{
+			name:    "wrong type",
+			jsonStr: `true`,
+			wantErr: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			iter := jsoniter.ConfigFastest.BorrowIterator([]byte(tt.jsonStr))
+			defer jsoniter.ConfigFastest.ReturnIterator(iter)
+			val := ReadFloat64(iter)
+			if tt.wantErr {
+				assert.Error(t, iter.Error)
+				return
+			}
+			assert.NoError(t, iter.Error)
+			assert.Equal(t, tt.want, val)
+		})
+	}
+}